Move Proxy and traps into their own files
